# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Estate360 API
  version: 1.0.0
  description: Real estate marketplace for Freetown, Sierra Leone.
paths:
...
  /api/v1/conversations/{id}:
    get:
      operationId: conversations_retrieve_2
      description: GET /conversations/{id}
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - conversations
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT